backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pymysql==1.1.1
alembic==1.15.2
openai==1.78.1
python-docx==1.1.2
python-multipart==0.0.20
pydantic-settings==2.8.1
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
email-validator>=2.0.0
anthropic>=0.52.0
google-generativeai>=0.8.0
